brands from frontend-api moved to packages
