+++ packages/framework/src/Model/Product/Brand/BrandRepository.php
@@ -5,7 +5,10 @@ declare(strict_types=1);
 namespace Shopsys\FrameworkBundle\Model\Product\Brand;
 
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Query\Expr\Join;
+use Doctrine\ORM\QueryBuilder;
 use Shopsys\FrameworkBundle\Component\Doctrine\OrderByCollationHelper;
+use Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\String\DatabaseSearching;
 use Shopsys\FrameworkBundle\Model\Product\Brand\Exception\BrandNotFoundException;
@@ -109,5 +112,57 @@ class BrandRepository
         $queryBuilder->orderBy(OrderByCollationHelper::createOrderByForLocale('b.name', $this->domain->getLocale()));
 
         return $queryBuilder->getQuery()->getResult();
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
+     * @return \Shopsys\FrameworkBundle\Model\Product\Brand\Brand[]
+     */
+    public function getAllWithDomainsAndTranslations(DomainConfig $domainConfig): array
+    {
+        return $this->getAllWithDomainsAndTranslationsQueryBuilder($domainConfig)
+            ->orderBy('b.name')
+            ->getQuery()->getResult();
+    }
+
+    /**
+     * @param int[] $brandIds
+     * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
+     * @return array<int, \Shopsys\FrameworkBundle\Model\Product\Brand\Brand|null>
+     */
+    public function getByIds(array $brandIds, DomainConfig $domainConfig): array
+    {
+        $result = $this->getAllWithDomainsAndTranslationsQueryBuilder($domainConfig)
+            ->andWhere('b.id IN (:brandIds)')
+            ->indexBy('b', 'b.id')
+            ->setParameter('brandIds', $brandIds)
+            ->getQuery()->getResult();
+
+        $brands = [];
+
+        foreach ($brandIds as $brandId) {
+            if (isset($result[$brandId])) {
+                $brands[] = $result[$brandId];
+            } else {
+                $brands[] = null;
+            }
+        }
+
+        return $brands;
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
+     * @return \Doctrine\ORM\QueryBuilder
+     */
+    protected function getAllWithDomainsAndTranslationsQueryBuilder(DomainConfig $domainConfig): QueryBuilder
+    {
+        return $this->em->createQueryBuilder()
+            ->select('b, bd, bt')
+            ->from(Brand::class, 'b')
+            ->join('b.domains', 'bd', Join::WITH, 'bd.domainId = :domainId')
+            ->join('b.translations', 'bt', Join::WITH, 'bt.locale = :locale')
+            ->setParameter('domainId', $domainConfig->getId())
+            ->setParameter('locale', $domainConfig->getLocale());
     }
 }
