+++ project-base/app/config/packages/overblog_dataloader.yaml
@@ -4,7 +4,7 @@ overblog_dataloader:
     loaders:
         brandsBatchLoader:
             alias: "brands_batch_loader"
-            batch_load_fn: "@App\\FrontendApi\\Model\\Brand\\BrandsBatchLoader:loadByIds"
+            batch_load_fn: "@Shopsys\\FrontendApiBundle\\Model\\Brand\\BrandsBatchLoader:loadByIds"
         categoriesBatchLoader:
             alias: "categories_batch_loader"
             batch_load_fn: "@Shopsys\\FrontendApiBundle\\Model\\Category\\CategoriesBatchLoader:loadByIds"
